+++ app/src/main/java/com/concordium/wallet/data/room/WalletDatabase.kt
@@ -1,16 +1,27 @@
 package com.concordium.wallet.data.room
 
 import android.content.Context
-import androidx.room.*
-import com.concordium.wallet.data.room.WalletDatabase.Companion.VERSION_NUMBER
+import androidx.room.AutoMigration
+import androidx.room.Database
+import androidx.room.Room
+import androidx.room.RoomDatabase
+import androidx.room.TypeConverters
 import com.concordium.wallet.data.room.migrations.MIGRATION_3_4
 import com.concordium.wallet.data.room.migrations.MIGRATION_4_5
 import com.concordium.wallet.data.room.migrations.MIGRATION_5_6
+import com.concordium.wallet.data.room.migrations.MIGRATION_7_8
 import com.concordium.wallet.data.room.typeconverter.GlobalTypeConverters
 
 @Database(
-    entities = [Identity::class, Account::class, Transfer::class, Recipient::class, EncryptedAmount::class, AccountContract::class, ContractToken::class],
-    version = VERSION_NUMBER,
+    entities = [
+        Identity::class,
+        Account::class,
+        Transfer::class,
+        Recipient::class,
+        EncryptedAmount::class,
+        ContractToken::class,
+    ],
+    version = 8,
     exportSchema = true,
     autoMigrations = [
         AutoMigration(from = 6, to = 7),
@@ -24,13 +35,9 @@ public abstract class WalletDatabase : RoomDatabase() {
     abstract fun transferDao(): TransferDao
     abstract fun recipientDao(): RecipientDao
     abstract fun encryptedAmountDao(): EncryptedAmountDao
-    abstract fun accountContractDao(): AccountContractDao
     abstract fun contractTokenDao(): ContractTokenDao
 
     companion object {
-
-        const val VERSION_NUMBER = 7
-
         // Singleton prevents multiple instances of database opening at the same time.
         @Volatile
         private var INSTANCE: WalletDatabase? = null
@@ -52,6 +59,7 @@ public abstract class WalletDatabase : RoomDatabase() {
                         MIGRATION_3_4,
                         MIGRATION_4_5,
                         MIGRATION_5_6,
+                        MIGRATION_7_8,
                     )
                     .build()
                 INSTANCE = instance