+++ app/src/main/java/com/concordium/wallet/ui/cis2/TokenDetailsActivity.kt
@@ -13,6 +13,7 @@ import com.concordium.wallet.data.model.TransactionStatus
 import com.concordium.wallet.data.room.Account
 import com.concordium.wallet.data.util.CurrencyUtil
 import com.concordium.wallet.databinding.ActivityTokenDetailsBinding
+import com.concordium.wallet.extension.showSingle
 import com.concordium.wallet.ui.account.accountdetails.AccountDetailsActivity
 import com.concordium.wallet.ui.account.accountqrcode.AccountQRCodeActivity
 import com.concordium.wallet.ui.base.BaseActivity
@@ -29,7 +30,7 @@ class TokenDetailsActivity : BaseActivity(R.layout.activity_token_details) {
     companion object {
         const val ACCOUNT = "ACCOUNT"
         const val TOKEN = "TOKEN"
-        const val DELETED = "DELETED"
+        const val CHANGED = "CHANGED"
     }
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -68,7 +69,10 @@ class TokenDetailsActivity : BaseActivity(R.layout.activity_token_details) {
             val intent = Intent(this, SendTokenActivity::class.java)
             intent.putExtra(SendTokenActivity.ACCOUNT, viewModel.tokenData.account)
             intent.putExtra(SendTokenActivity.TOKEN, viewModel.tokenData.selectedToken)
-            intent.putExtra(SendTokenActivity.PARENT_ACTIVITY, AccountDetailsActivity::class.java.canonicalName)
+            intent.putExtra(
+                SendTokenActivity.PARENT_ACTIVITY,
+                AccountDetailsActivity::class.java.canonicalName
+            )
             startActivityForResultAndHistoryCheck(intent)
         }
         binding.includeButtons.receive.setOnClickListener {
@@ -93,7 +97,25 @@ class TokenDetailsActivity : BaseActivity(R.layout.activity_token_details) {
                 setTicker(tokenMetadata)
                 setDecimals(tokenMetadata)
             }
+            if (token.isNewlyReceived) {
+                handleNewlyReceivedToken(token)
+            }
         }
+    }
+
+    private fun handleNewlyReceivedToken(token: Token) {
+        showNewlyReceivedNotice(
+            tokenName = token.name ?: token.symbol
+        )
+        viewModel.unmarkNewlyReceivedToken(
+            viewModel.tokenData.account!!.address,
+            viewModel.tokenData.selectedToken!!.contractIndex,
+            viewModel.tokenData.selectedToken!!.id
+        )
+        setResult(
+            Activity.RESULT_OK,
+            Intent().putExtra(CHANGED, true)
+        )
     }
 
     private fun showDeleteDialog() {
@@ -107,9 +129,10 @@ class TokenDetailsActivity : BaseActivity(R.layout.activity_token_details) {
                 viewModel.tokenData.selectedToken!!.contractIndex,
                 viewModel.tokenData.selectedToken!!.id
             )
-            val intent = Intent()
-            intent.putExtra(DELETED, true)
-            setResult(Activity.RESULT_OK, intent)
+            setResult(
+                Activity.RESULT_OK,
+                Intent().putExtra(CHANGED, true)
+            )
             finish()
         }
         builder.setNegativeButton(getString(R.string.cis_delete_dialog_cancel)) { dialog, _ ->
@@ -206,9 +229,27 @@ class TokenDetailsActivity : BaseActivity(R.layout.activity_token_details) {
         }
     }
 
+    private fun showNewlyReceivedNotice(tokenName: String) {
+        NewlyReceivedTokenNoticeDialog.newInstance(
+            NewlyReceivedTokenNoticeDialog.getBundle(
+                tokenName = tokenName,
+            )
+        ).showSingle(supportFragmentManager, NewlyReceivedTokenNoticeDialog.TAG)
+    }
+
     private fun initObservers() {
         viewModel.waiting.observe(this) { waiting ->
             showWaiting(waiting)
+        }
+
+        supportFragmentManager.setFragmentResultListener(
+            NewlyReceivedTokenNoticeDialog.ACTION_REQUEST,
+            this,
+        ) { _, bundle ->
+            val isKeepingToken = NewlyReceivedTokenNoticeDialog.getResult(bundle)
+            if (!isKeepingToken) {
+                showDeleteDialog()
+            }
         }
     }