+++ app/src/main/java/com/concordium/wallet/ui/cis2/defaults/DefaultFungibleTokensManager.kt
@@ -1,14 +1,11 @@
 package com.concordium.wallet.ui.cis2.defaults
 
-import com.concordium.wallet.data.AccountContractRepository
 import com.concordium.wallet.data.ContractTokensRepository
-import com.concordium.wallet.data.room.AccountContract
 import com.concordium.wallet.util.Log
 
 class DefaultFungibleTokensManager(
     private val defaults: Collection<DefaultFungibleToken>,
     private val contractTokensRepository: ContractTokensRepository,
-    private val accountContractRepository: AccountContractRepository,
 ) {
     suspend fun addForAccount(accountAddress: String) {
         Log.d(
@@ -18,27 +15,9 @@ class DefaultFungibleTokensManager(
         )
 
         defaults.forEach { tokenToAdd ->
-            val accountContract =
-                accountContractRepository.find(
-                    accountAddress = accountAddress,
-                    contractIndex = tokenToAdd.contractIndex,
-                )
-            if (accountContract == null) {
-                accountContractRepository.insert(
-                    AccountContract(
-                        id = 0,
-                        accountAddress = accountAddress,
-                        contractIndex = tokenToAdd.contractIndex,
-                    )
-                )
-
-                Log.d(
-                    "created_account_contract:" +
-                            "\ncontractIndex=${tokenToAdd.contractIndex}"
-                )
-            }
-
-            val contractToken = tokenToAdd.toContractToken(accountAddress)
+            val contractToken = tokenToAdd.toContractToken(
+                accountAddress = accountAddress,
+            )
             contractTokensRepository.insert(contractToken)
         }