+++ app/src/main/java/com/concordium/wallet/ui/passphrase/recoverprocess/RecoverProcessViewModel.kt
@@ -8,7 +8,7 @@ import androidx.lifecycle.viewModelScope
 import com.concordium.wallet.App
 import com.concordium.wallet.AppConfig
 import com.concordium.wallet.core.backend.BackendRequest
-import com.concordium.wallet.data.AccountContractRepository
+import com.concordium.wallet.core.notifications.UpdateNotificationsSubscriptionUseCase
 import com.concordium.wallet.data.AccountRepository
 import com.concordium.wallet.data.ContractTokensRepository
 import com.concordium.wallet.data.IdentityRepository
@@ -79,6 +79,9 @@ class RecoverProcessViewModel(application: Application) : AndroidViewModel(appli
     private var identityProviders: ArrayList<IdentityProvider>? = null
     private val identityGapMutex = Mutex()
     private val accountGapMutex = Mutex()
+    private val updateNotificationsSubscriptionUseCase by lazy {
+        UpdateNotificationsSubscriptionUseCase(application)
+    }
 
     val statusChanged: MutableLiveData<Int> by lazy { MutableLiveData<Int>() }
     val waiting: MutableLiveData<Boolean> by lazy { MutableLiveData<Boolean>() }
@@ -95,10 +98,8 @@ class RecoverProcessViewModel(application: Application) : AndroidViewModel(appli
         val recipientDao = WalletDatabase.getDatabase(application).recipientDao()
         recipientRepository = RecipientRepository(recipientDao)
         val contractTokenDao = WalletDatabase.getDatabase(application).contractTokenDao()
-        val accountContractDao = WalletDatabase.getDatabase(application).accountContractDao()
         val defaultTokensManagerFactory = DefaultTokensManagerFactory(
             contractTokensRepository = ContractTokensRepository(contractTokenDao),
-            accountContractRepository = AccountContractRepository(accountContractDao),
         )
         defaultFungibleTokensManager = defaultTokensManagerFactory.getDefaultFungibleTokensManager()
     }
@@ -354,6 +355,7 @@ class RecoverProcessViewModel(application: Application) : AndroidViewModel(appli
                         recipientRepository.insert(Recipient(account))
                     }
                     defaultFungibleTokensManager.addForAccount(account.address)
+                    updateNotificationsSubscriptionUseCase()
 
                     val iWithAFound =
                         identitiesWithAccountsFound.firstOrNull { it.identity.identityProviderId == identity.identityProviderId && it.identity.identityIndex == identity.identityIndex }